requirements.txt
google-genai
python-dotenv
python-frontmatter
streamlit
pandas
pypandoc
python-docx
openpyxl
requests
watchdog
openai
anthropic
